moving files and small updates
