+++ src/evaluate_images.py
