+++ src/generate_class_images.py
