+++ src/generate_test_images_v1.py
