+++ src/old2/generate_synthetic_images.py
